Change: add support for Chrome v49*
* Last version supported on Windows XP and Mac OS X Snow Leopard
