+++ .babelrc.js
@@ -1,0 +1,14 @@
+const modeDevelopment = process.env.NODE_ENV === 'development';
+
+module.exports = {
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        debug: modeDevelopment,
+        useBuiltIns: 'usage',
+        corejs: 3,
+      },
+    ],
+  ],
+};
