+++ .browserslistrc
@@ -1,0 +1,2 @@
+# Chrome v49 is the last version supported on Windows XP and Mac OS X Snow Leopard
+Chrome > 48
