+++ src/manifest.json
@@ -32,7 +32,6 @@
       "global/js/i18next/i18next.min.js",
       "global/js/i18next/i18nextBrowserLanguageDetector.min.js",
       "global/js/i18next/i18nextXHRBackend.js",
-      "global/js/Array.from.polyfill.js",
       "global/js/i18n.js",
       "global/js/context-menus.js",
       "global/js/background.js",
@@ -76,7 +75,6 @@
         "global/js/i18next/i18next.min.js",
         "global/js/i18next/i18nextBrowserLanguageDetector.min.js",
         "global/js/i18next/i18nextXHRBackend.js",
-        "global/js/Array.from.polyfill.js",
         "global/js/i18n.js",
         "global/js/punycode.min.js",
         "shared/custom-elements/custom-elements.js",
