+++ static/browser-action/index.html
@@ -14,7 +14,6 @@
     <script src="/global/js/i18next/i18next.min.js"></script>
     <script src="/global/js/i18next/i18nextBrowserLanguageDetector.min.js"></script>
     <script src="/global/js/i18next/i18nextXHRBackend.js"></script>
-    <script src="/global/js/Array.from.polyfill.js"></script>
     <script src="/global/js/i18n.js"></script>
     <script src="/global/js/page.js"></script>
     <script src="/global/js/punycode.min.js"></script>
