+++ static/options/index.html
@@ -259,11 +259,10 @@
     <script src="/global/js/i18next/i18next.min.js"></script>
     <script src="/global/js/i18next/i18nextBrowserLanguageDetector.min.js"></script>
     <script src="/global/js/i18next/i18nextXHRBackend.js"></script>
-    <script src="/global/js/Array.from.polyfill.js"></script>
     <script src="/global/js/i18n.js"></script>
     <script src="/global/js/page-v2.js"></script>
     <script src="/global/js/context-menus.js"></script>
     <script src="/global/js/incentive.js"></script>
-    <script src="/options/js/options.js"></script>
+    <script src="/options/options.js"></script>
   </body>
 </html>
